+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 ### Quality
 
 - Remove `EncryptionService`. (#127)
+- Adapt `PostComputeApp` and `Web2ResultService` to embrace API breaking changes in iexec-common. (#127)
 
 ### Dependency Upgrades